# docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"HttpResponse_string_": {
				"properties": {
					"payload": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticateRequestBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TComment": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"post_id": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"post_id",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_TComment_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/TComment"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"USER_ROLES": {
				"enum": [
					"NORMAL",
					"ADMIN"
				],
				"type": "string"
			},
			"TokenPayload": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/USER_ROLES"
					}
				},
				"required": [
					"id",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					},
					"requestPostId": {
						"type": "string"
					}
				},
				"required": [
					"requestContent",
					"requestUser",
					"requestPostId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TPost": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_TPost_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/TPost"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestContent",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteCommentRequest": {
				"properties": {
					"requestId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeletePostRequest": {
				"properties": {
					"requestId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FetchCommentsOutput": {
				"properties": {
					"creator": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"no updates"
								]
							}
						]
					},
					"createdAt": {
						"type": "string"
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"creator",
					"updatedAt",
					"createdAt",
					"dislikes",
					"likes",
					"content",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_FetchCommentsOutput-Array_": {
				"properties": {
					"payload": {
						"items": {
							"$ref": "#/components/schemas/FetchCommentsOutput"
						},
						"type": "array"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FetchPostsOutput": {
				"properties": {
					"creator": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"creator",
					"createdAt",
					"dislikes",
					"likes",
					"content",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_FetchPostsOutput-Array_": {
				"properties": {
					"payload": {
						"items": {
							"$ref": "#/components/schemas/FetchPostsOutput"
						},
						"type": "array"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPost": {
				"properties": {
					"comments": {
						"items": {
							"$ref": "#/components/schemas/FetchCommentsOutput"
						},
						"type": "array"
					},
					"creator": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"no updates"
								]
							}
						]
					},
					"createdAt": {
						"type": "string"
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"comments",
					"creator",
					"updatedAt",
					"createdAt",
					"content",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_GetPost_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/GetPost"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserOutput": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_UserOutput_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/UserOutput"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LikeDislikeRequest": {
				"properties": {
					"requestLike": {
						"type": "boolean"
					},
					"requestContentId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestLike",
					"requestContentId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequestBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"post_id": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"post_id",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_Comment_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/Comment"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCommentRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestCommentId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestContent",
					"requestCommentId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Post": {
				"properties": {
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"dislikes": {
						"type": "number",
						"format": "double"
					},
					"likes": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"creator_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"content",
					"creator_id",
					"id"
				],
				"type": "object"
			},
			"HttpResponse_Post_": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/Post"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payload",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePostRequestContent": {
				"properties": {
					"requestContent": {
						"type": "string"
					},
					"requestId": {
						"type": "string"
					},
					"requestUser": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"requestContent",
					"requestId",
					"requestUser"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "labook",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/authenticate": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_string_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticateRequestBody"
							}
						}
					}
				}
			}
		},
		"/posts/{id}/comments": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_TComment_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequestContent"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_FetchCommentsOutput-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_TPost_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostRequestContent"
							}
						}
					}
				}
			}
		},
		"/comments/{id}": {
			"delete": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteCommentRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_Comment_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCommentRequestContent"
							}
						}
					}
				}
			}
		},
		"/posts/{id}": {
			"delete": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletePostRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_GetPost_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_Post_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostRequestContent"
							}
						}
					}
				}
			}
		},
		"/posts/page/{page}": {
			"get": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_FetchPostsOutput-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_UserOutput_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/{id}/like": {
			"put": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LikeDislikeRequest"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpResponse_string_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequestBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3009/v1"
		}
	]
}